add from file

+++ src/Sample/SamplingMethod.h
@@ -10,7 +10,7 @@ class SamplingMethod
 {
     
 public:
-    virtual void create(boost::shared_ptr<SampleIndividual>& ind, int i) = 0;
+    virtual void create(boost::shared_ptr<SampleIndividual>& ind, size_t i) = 0;
 };
 
 #endif
